Fixed bindings for editing; fixed some styles for content

+++ CPI/ClientApp/src/app/patient/user-page/user-page.component.ts
@@ -2,9 +2,13 @@ import { Time } from '@angular/common';
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Template } from '@angular/compiler/src/render3/r3_ast';
 import { Component, Inject } from '@angular/core';
+import { MatDialog } from '@angular/material';
 import { ActivatedRoute, Router } from '@angular/router';
+import { dialogConfig, roles } from '../../../globals';
+import { DeleteDialogComponent } from '../../admin/delete-dialog/delete-dialog.component';
 import { Speciality } from '../../admin/doctor/doctors-list/doctors-list.component';
 import { Patient } from '../../admin/patient/patients-list/patients-list.component';
+import { PassData } from '../../admin/user/add-user/add-user.component';
 
 @Component({
   selector: 'app-user-page',
@@ -17,18 +21,27 @@ export class UserPageComponent {
   public login: string;
   public coupons: Coupon[] = [];
   public doctor: Doctor = new Doctor("", "", "", 0, "", new Speciality("", ""));
+  public isUser: boolean = true;
 
-  constructor(private router: Router, private activateRoute: ActivatedRoute, private http: HttpClient, @Inject('BASE_URL') private baseUrl: string) {
+  constructor(private router: Router, private activateRoute: ActivatedRoute, private http: HttpClient, @Inject('BASE_URL') private baseUrl: string,
+    private dialog: MatDialog  ) {
     this.login = activateRoute.snapshot.params['login'];
     this.getUser(this.login);
     this.getCoupons();
+  }
+
+  checkStatus() {
+    this.http.get<PassData>(this.baseUrl + 'passdata/' + this.login).subscribe(result => {
+      result.status == roles[0] ? this.isUser : this.isUser = !this.isUser;
+      console.log(this.isUser);
+      console.log(result.status);
+    }, error => console.error(error));
   }
 
   getUser(login: string) {
     this.http.get<Patient>(this.baseUrl + 'patient/' + login).subscribe(result => {
       this.user = result;
     }, error => console.error(error));
-    console.log(this.user);
   }
 
   getCoupons() {
@@ -49,12 +62,17 @@ export class UserPageComponent {
   }
 
   deleteCoupon(coupon_id: number) {
-    this.http.delete(this.baseUrl + 'coupon', { params: new HttpParams().set('coupon_id', coupon_id.toString()) })
-      .subscribe(
-        (data: any) => {
-          this.getCoupons();
-        },
-        error => console.log(error));
+    const dialogRef = this.dialog.open(DeleteDialogComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe(data => {
+      if (data != null) {
+        this.http.delete(this.baseUrl + 'coupon', { params: new HttpParams().set('coupon_id', coupon_id.toString()) })
+          .subscribe(
+            (data: any) => {
+              this.getCoupons();
+            },
+            error => console.log(error));
+      }
+    });
   }
 
   getDoctorByCoup(id: number) {
